requirements.txt
# Dev dependencies
pytest==6.2.3
pytest-mock==3.5.1
requests-mock==1.9.2

# Dependencies
beautifulsoup4==4.8.2
elasticsearch==7.8.0
elasticsearch-dsl==7.2.1
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
lxml==4.6.3
pandas==0.25.3
python-keystoneclient==4.0.0
python-swiftclient==3.9.0
python-dateutil~=2.8.1
redis==3.5.3
retry~=0.9.2
requests==2.19.1
rq==1.9.0
tokenizers==0.10.1
